# lib/api-zod/src/generated/types/activityLog.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Onlinelandregistry.uk API
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityLog {
  id: number;
  orderId: number;
  action: string;
  /** @nullable */
  detail?: string | null;
  /** @nullable */
  author?: string | null;
  createdAt: string;
}