# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Onlinelandregistry.uk API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderCountry } from './orderCountry';
import type { OrderDeliveryType } from './orderDeliveryType';
import type { OrderNotificationType } from './orderNotificationType';
import type { OrderStatus } from './orderStatus';
import type { OrderTrackingType } from './orderTrackingType';

export interface Order {
  id: number;
  orderNumber: string;
  status: OrderStatus;
  serviceId: number;
  serviceName: string;
  customerName: string;
  customerEmail: string;
  /** @nullable */
  customerPhone?: string | null;
  propertyCount?: number;
  country?: OrderCountry;
  /** @nullable */
  tenure?: string | null;
  /** @nullable */
  titleNumber?: string | null;
  /** @nullable */
  postcode?: string | null;
  /** @nullable */
  propertyAddress?: string | null;
  addons?: string[];
  trackingType?: OrderTrackingType;
  deliveryType?: OrderDeliveryType;
  notificationType?: OrderNotificationType;
  documentFee?: number;
  serviceFee?: number;
  vatAmount?: number;
  totalAmount: number;
  /** @nullable */
  stripeSessionId?: string | null;
  /** @nullable */
  paidAt?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  staffNotes?: string | null;
  createdAt: string;
  updatedAt?: string;
}