# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Onlinelandregistry.uk API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderInputCountry } from './orderInputCountry';
import type { OrderInputDeliveryType } from './orderInputDeliveryType';
import type { OrderInputNotificationType } from './orderInputNotificationType';
import type { OrderInputTrackingType } from './orderInputTrackingType';

export interface OrderInput {
  serviceId: number;
  customerTitle?: string;
  customerName: string;
  customerEmail: string;
  customerEmailConfirm?: string;
  customerPhone?: string;
  customerAddress?: string;
  propertyCount: number;
  country: OrderInputCountry;
  tenure?: string;
  titleNumber?: string;
  postcode?: string;
  propertyAddress?: string;
  /** @nullable */
  lat?: number | null;
  /** @nullable */
  lng?: number | null;
  addons: string[];
  trackingType: OrderInputTrackingType;
  deliveryType: OrderInputDeliveryType;
  notificationType: OrderInputNotificationType;
  agreedToWaiveCancel?: boolean;
}