# lib/api-zod/src/generated/types/orderList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Onlinelandregistry.uk API
 * OpenAPI spec version: 0.1.0
 */
import type { Order } from './order';

export interface OrderList {
  orders: Order[];
  total: number;
}