# lib/api-zod/src/generated/types/priceBreakdown.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Onlinelandregistry.uk API
 * OpenAPI spec version: 0.1.0
 */
import type { LineItem } from './lineItem';

export interface PriceBreakdown {
  documentFee: number;
  serviceFee: number;
  vatAmount: number;
  totalAmount: number;
  lineItems: LineItem[];
}