# lib/api-zod/src/generated/types/serviceCount.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Onlinelandregistry.uk API
 * OpenAPI spec version: 0.1.0
 */

export interface ServiceCount {
  serviceName: string;
  count: number;
  revenue: number;
}